feat(app): restructure navigators and screens
